refactor: fix review comments for tests and improve package structure
